# Speed up `PandasDataset` further